New changes in triggers and CREATE
Two new triggers added. Further discussion needed for the last trigger which makes sure that the supervisor is the last researcher of a project to be deleted.  
Deliverable trigger has been deleted and deliverable table has been edited on the foreign key proj_id from ON DELETE RESTRICT to ON DELETE CASCADE.
Now when we delete a project, that project's deliverables get deleted as well.